set up the assets folder to allow debugging of an errored workout on a separate device by loading assets/debug/saved_error_state.json file
