+++ .gitignore
@@ -46,5 +46,5 @@ app.*.map.json
 *db
 *db-journal
 
-.debug-error-workouts
+assets/debug/saved_error_state.json
